Dashboard mejorado
