--- Client/src/Components/Views/AdminTableUsers/AdminTableUsers.module.css
@@ -1,11 +1,0 @@
-/* Estilos para el paginado */
-.customPagination {
-  background-color: #DFE2E2;
-  display: flex;
-	justify-content: end;
-  align-items: center;
-}
-.fila:hover{
-  background-color: #DDE6ED;
-  cursor: pointer;
-}
