+++ Client/src/Components/Views/Dashboard/AdminTableProperties/AdminTableProperties.module.css
@@ -1,0 +1,72 @@
+/* Estilos para el paginado */
+.customPagination {
+  background-color: #DFE2E2;
+  display: flex;
+	justify-content: end;
+  align-items: center;
+}
+.fila:hover{
+  background-color: #DDE6ED;
+  cursor: pointer;
+}
+
+
+
+/*Modal*/
+.modalContainer {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: 9999;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.fade{
+  padding: 0;
+  width: 80vw;
+  height: 80vh;
+  text-align: center;
+}
+
+.cardHeader{
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
+
+.cardBody{
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  align-items: center;
+}
+
+.cardImage{
+  max-width: 50vw;
+}
+
+.cardData{
+  font-size: 20px;
+}
+
+.cardInfo{
+  display: flex;
+  flex-direction: column;
+  gap: 2vw;
+  text-align: left;
+}
+
+.cardTitle{
+  font-size: 30px;
+}
+
+.propertyCard{
+  background-color: white;
+  width: 80vw;
+  height: 80vh;
+  text-align: center;
+}
